jegliche IDs in long umgeschrieben

+++ SopraMMS/src/management/CourseEntry.java
@@ -12,21 +12,21 @@ public class CourseEntry extends Entry {
 
 
 	public CourseEntry(int version, String date, boolean classification,
-			boolean approvalstatus, boolean declined, int entryID, String title, List<String> courses) {
+			boolean approvalstatus, boolean declined, long entryID, String title, List<String> courses) {
 		super(version, date, classification, approvalstatus, declined, entryID, title);
 		this.courses = courses;
 	}
 	
 	
 	public CourseEntry(int version, String date, boolean classification,
-			boolean approvalstatus, boolean declined, int entryID, String title, String course) {
+			boolean approvalstatus, boolean declined, long entryID, String title, String course) {
 		super(version, date, classification, approvalstatus, declined, entryID, title);
 		courses.add(course);
 	}
 	
 	
 	public CourseEntry(int version, String date, boolean classification,
-			boolean approvalstatus, boolean declined, int entryID, String title) {
+			boolean approvalstatus, boolean declined, long entryID, String title) {
 		super(version, date, classification, approvalstatus, declined, entryID, title);
 		courses = new LinkedList<String>();
 	}
